requirements.txt
streamlit
langchain
langchain-groq
langchain-community
faiss-cpu
pypdf
python-dotenv
sentence-transformers